Refactor to update state when cache or SLURM changes
Before this commit the new stayrtr state was only recalculated
when the validated cache changed.

This commit refactors the update loop and separates the actual
update from pulling the updated cache. The update is triggered
when the SLURM file changes _or_ when the validated cache changes.

Should fix https://github.com/cloudflare/gortr/issues/95 and make
further changes that trigger an update when an update in the cache
has expired easier.