Avoid sending stale data, properly handle errRPKIJsonFileTooOld
Both updateFromNewState() and reloadFromCurrentState() can return
errRPKIJsonFileTooOld.

However the RTR clearing code that would withdraw all VRPs is only in the call
site of updateFromNewState(), which is not called when the JSON file is stuck,
because CacheUpdated is false (as updateFile() returns false, IdenticalFile).

That is why with a hung RPKI validator and logging turned up to info we see:

File /var/lib/rpki-client/json is identical to the previous version
RPKI JSON file is older than 24 hours: 2025-09-01 16:29:30 +0000 UTC
Error updating state: RPKI JSON file is older than 24 hours

But no clearing of the VRPs occurs, because the reloadFromCurrentState()
caller doesn't handle errRPKIJsonFileTooOld specifically and just prints
an error.

Move errRPKIJsonFileTooOld handling into it's own function and call the
handler from both call sites. Although I'm not sure about the use-case for
updateFromNewState(), if it is returning errRPKIJsonFileTooOld I want to
handle it properly. I also want to avoid duplicate handling code, because
that will bite us again in the future if this code ever changes.